SAAS-108: Adding user tests (#42)
* SAAS-108: Making items test not affect users on prod db

* SAAS-108: Adding tests for fetching users

* SAAS-108: Moved update function to users

* SAAS-108: Added tests for updating profile

* SAAS-108: Added tests for deleting a user

* SAAS-108: Finished tests for users

* SAAS-108: Fixed bug in item tests

* SAAS-108: Chnaged id to user id

---------

+++ tests/items.test.ts
@@ -1,21 +1,9 @@
 import { expect, test, describe, spyOn, beforeAll, afterAll } from "bun:test";
 import {
-  generateUser,
-  checkUser,
   register,
-  refresh,
   login,
-  logout,
-  logoutAll,
-  forgotPassword,
-  resetPassword,
 } from "../src/application/user_application";
-import { afterEach, beforeEach, mock } from "node:test";
 import pg, { redis } from "../src/utils/db";
-import { createHash } from "node:crypto";
-import { verifyRefreshToken } from "../src/utils/jwt_config";
-import { getMaxListeners } from "node:cluster";
-import { sleep } from "bun";
 import {
   getItemsById,
   getAllItems,
@@ -32,10 +20,9 @@ let sellerId: string | null = null;
 let sellerId2: string | null = null;
 
 beforeAll(async () => {
-  // register users
-  await pg`delete from refresh_tokens`;
-  await pg`delete from items`;
-  await pg`delete from users`;
+  await pg`delete from items where item_id in (${itemId1}, ${itemId2}, ${itemId3})`;
+  await pg`delete from refresh_tokens where user_id in (select user_id from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com'))`;
+  await pg`delete from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com')`;
 
   const registerReq = generateRequest(
     "http://localhost/auth/register",
@@ -85,18 +72,16 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
-  // delete all registered users
-  await pg`delete from refresh_tokens`;
-  await pg`delete from items`;
-  await pg`delete from users`;
+  await pg`delete from items where item_id in (${itemId1}, ${itemId2}, ${itemId3})`;
+  await pg`delete from refresh_tokens where user_id in (select user_id from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com'))`;
+  await pg`delete from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com')`;
 });
 
 describe("Getting items tests", () => {
   beforeAll(async () => {
-    // register users
-    await pg`delete from refresh_tokens`;
-    await pg`delete from items`;
-    await pg`delete from users`;
+    await pg`delete from items where item_id in (${itemId1}, ${itemId2}, ${itemId3})`;
+    await pg`delete from refresh_tokens where user_id in (select user_id from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com'))`;
+    await pg`delete from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com')`;
 
     const registerReq = generateRequest(
       "http://localhost/auth/register",
@@ -146,10 +131,9 @@ describe("Getting items tests", () => {
   });
 
   afterAll(async () => {
-    // delete all registered users
-    await pg`delete from refresh_tokens`;
-    await pg`delete from items`;
-    await pg`delete from users`;
+    await pg`delete from items where item_id in (${itemId1}, ${itemId2}, ${itemId3})`;
+    await pg`delete from refresh_tokens where user_id in (select user_id from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com'))`;
+    await pg`delete from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com')`;
   });
 
   test("Getting item by item id", async () => {
@@ -344,9 +328,9 @@ describe("Getting items tests", () => {
 describe("Update item tests", () => {
   beforeAll(async () => {
     // register users
-    await pg`delete from refresh_tokens`;
-    await pg`delete from items`;
-    await pg`delete from users`;
+    await pg`delete from items where item_id in (${itemId1}, ${itemId2}, ${itemId3})`;
+    await pg`delete from refresh_tokens where user_id in (select user_id from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com'))`;
+    await pg`delete from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com')`;
 
     const registerReq = generateRequest(
       "http://localhost/auth/register",
@@ -396,10 +380,9 @@ describe("Update item tests", () => {
   });
 
   afterAll(async () => {
-    // delete all registered users
-    await pg`delete from refresh_tokens`;
-    await pg`delete from items`;
-    await pg`delete from users`;
+    await pg`delete from items where item_id in (${itemId1}, ${itemId2}, ${itemId3})`;
+    await pg`delete from refresh_tokens where user_id in (select user_id from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com'))`;
+    await pg`delete from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com')`;
   });
 
   test("No items were provided", async () => {
@@ -496,9 +479,9 @@ describe("Update item tests", () => {
 describe("Deleting item tests", () => {
   beforeAll(async () => {
     // register users
-    await pg`delete from refresh_tokens`;
-    await pg`delete from items`;
-    await pg`delete from users`;
+    await pg`delete from items where item_id in (${itemId1}, ${itemId2}, ${itemId3})`;
+    await pg`delete from refresh_tokens where user_id in (select user_id from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com'))`;
+    await pg`delete from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com')`;
 
     const registerReq = generateRequest(
       "http://localhost/auth/register",
@@ -549,9 +532,9 @@ describe("Deleting item tests", () => {
 
   afterAll(async () => {
     // delete all registered users
-    await pg`delete from refresh_tokens`;
-    await pg`delete from items`;
-    await pg`delete from users`;
+    await pg`delete from items where item_id in (${itemId1}, ${itemId2}, ${itemId3})`;
+    await pg`delete from refresh_tokens where user_id in (select user_id from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com'))`;
+    await pg`delete from users where email in ('jasonli1234@gmail.com', 'jasonli8909@gmail.com')`;
   });
 
   test("Entry does not exist", async () => {
@@ -591,6 +574,7 @@ describe("Deleting item tests", () => {
       accessToken,
     );
 
+    const query2 = await pg`select * from items`;
     const response = await deleteItem(request2);
     const body = await response.json();
 
@@ -599,6 +583,6 @@ describe("Deleting item tests", () => {
     expect(body.response).not.toBe(undefined);
 
     const query = await pg`select * from items`;
-    expect(query.length).toBe(2);
+    expect(query.length).not.toBe(query2.length);
   });
 });
